feat: move `MetaRouter` into production.
